Separate compressed file types into separate files.
Cleaner this way.
