+++ pcsx2/CDVD/IsoFileFormats.h
@@ -18,6 +18,7 @@
 #include "CDVD.h"
 #include "wx/wfstream.h"
 #include "AsyncFileReader.h"
+#include "CompressedFileReader.h"
 
 enum isoType
 {
